llava_service/requirements.txt
accelerate==1.11.0
bitsandbytes==0.48.1
fastapi==0.116.1
huggingface-hub==0.35.3
pillow==12.1.0
python-multipart==0.0.20
sentencepiece==0.2.1
transformers==4.57.1
uvicorn==0.37.0
